+++ src/app/buttonGroup/buttonGroup.sample.ts
@@ -1,6 +1,14 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
 import { NgFor } from '@angular/common';
-import { ButtonGroupAlignment, IgxButtonDirective, IgxButtonGroupComponent, IgxIconComponent, IgxLayoutDirective } from 'igniteui-angular';
+import { FormsModule } from '@angular/forms';
+import {
+	ButtonGroupAlignment,
+	IgxButtonDirective,
+	IgxButtonGroupComponent,
+	IgxIconComponent,
+	IgxLayoutDirective,
+	IgxSwitchComponent,
+} from 'igniteui-angular';
 
 
 interface IButton {
@@ -40,7 +48,7 @@ class Button {
     selector: 'app-buttongroup-sample',
     templateUrl: 'buttonGroup.sample.html',
     standalone: true,
-    imports: [IgxButtonGroupComponent, IgxButtonDirective, IgxIconComponent, NgFor, IgxLayoutDirective]
+	imports: [FormsModule, IgxButtonGroupComponent, IgxButtonDirective, IgxIconComponent, NgFor, IgxLayoutDirective, IgxSwitchComponent],
 })
 
 export class ButtonGroupSampleComponent implements OnInit {
@@ -53,8 +61,9 @@ export class ButtonGroupSampleComponent implements OnInit {
     public fontOptions: Button[];
     public cities: Button[];
     public borders: Button[];
-    public buttons = ['One', 'Two', 'Three', 'Four'];
+    public buttons = ['One', 'Two', 'Three, long text item', 'Four'];
     public selectedIndex = 1;
+	public isJustify = false;
 
     constructor() { }