+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes for each version of this project will be documented in this 
 ## 16.0.0
 
 ### General
+- `IgxButtonGroup`
+  - `igx-buttongroup` component has new input property `justifyItems`, the property makes all items inside the group equal in size, this is helpful when the group is in horizontal mode.
+
+  ```HTML
+  <igx-buttongroup justifyItems>
+      ...
+  </igx-buttongroup>
+   ```
+
 - All Ignite UI for Angular components are now exported as `standalone` components. The library still exports `NgModules`, which have been preserved for backward compatibility, but they no longer declare any of the Ignite UI for Angular components, instead they just import and export the `standalone` components. The `standalone` components are still in a preview stage. Some utility directive exports may change in the future and may be missing from the documentation in the initial release, hence the `preview` state of the feature.
 
   Now you can do: